- Fixed a bug with the dawrvish translation guide when not using museumsanity
